requirements.txt
fabric
boto
mako
lxml
paramiko
pycrypto
requests==2.9.1
azure-mgmt-commerce
azure-common
azure-mgmt-resource